refactor: fix indent and isolate responsible of translate
